+++ src/helpers.ts
@@ -9,6 +9,11 @@ export function replacePartAtPos(
 	return before + newText + after;
 }
 
+/**
+ * Returns a random integer between (and including) `min` and `max`.
+ * @param min The lower end of the possible range.
+ * @param max The upper end of the possible range.
+ */
 export function getRandomInt(min: number, max: number): number {
 	return Math.floor(Math.random() * (max - min)) + min;
 }
@@ -20,4 +25,34 @@ export function padNumber(num: number, width: number, z: string = '0'): string {
 
 export function getFilenameFromPath(path: string): string {
 	return path.replace(/^.*[\\\/]/, '');
+}
+
+/**
+ * Capitalizes the string's first character.
+ * @param text The string to be capitalized.
+ * @param all If true, _all_ words in the string will be capitalized.
+ */
+export function capitalize(text: string, all: boolean = false): string {
+	if (all) {
+		return text.replace(/\w\S*/g, word => capitalize(word));
+	}
+	return text.replace(/^\w/, c => c.toUpperCase());
+}
+
+const arrayMoveMutate = (array: any[], from: number, to: number) => {
+	const startIndex = to < 0 ? array.length + to : to;
+	const item = array.splice(from, 1)[0];
+	array.splice(startIndex, 0, item);
+};
+/**
+ * Move an array item to a different position
+ * @param array
+ * @param from Index of item to move. If negative, it will begin that many elements from the end.
+ * @param to Index of where to move the item. If negative, it will begin that many elements from the end.
+ * @source [[npm] array-move](https://www.npmjs.com/package/array-move)
+ */
+export function arrayMove(array: any[], from: number, to: number): any[] {
+	array = array.slice();
+	arrayMoveMutate(array, from, to);
+	return array;
 }